[ClimbFight#9] 🚚 colors props를 별도 파일로 분리
