+++ app/styles/light_theme.js
@@ -1,0 +1,19 @@
+const colors = {
+    '--grey-1': '#343a40',
+    '--grey-2': '#495057',
+    '--grey-3': '#868e96',
+    '--grey-4': '#adb5bd',
+    '--grey-5': '#ced4da',
+    '--white-2': '#f1f3f5',
+    '--white-1': '#f8f9fa',
+    '--white': '#ffffff',
+    '--marine': '#07cfcc',
+    '--pale-marine': '#ccfffe',
+    '--red': '#ff4523',
+}
+
+const light_theme = {
+    colors,
+}
+
+export default light_theme
