+++ app/styles/theme.js
@@ -33,20 +33,6 @@ const margins = {
     xxxxl: addPx(32),
 }
 
-const colors = {
-    '--grey-1': '#343a40',
-    '--grey-2': '#495057',
-    '--grey-3': '#868e96',
-    '--grey-4': '#adb5bd',
-    '--grey-5': '#ced4da',
-    '--white-2': '#f1f3f5',
-    '--white-1': '#f8f9fa',
-    '--white': '#ffffff',
-    '--marine': '#07cfcc',
-    '--pale-marine': '#ccfffe',
-    '--red': '#ff4523',
-}
-
 const fonts = {
     ko_light: `
       font-family: AppleSDGothicNeoL;
@@ -91,7 +77,6 @@ const verticalInterval = {
 
 const theme = {
     fontSizes,
-    colors,
     fonts,
     paddings,
     margins,
